[refactoring] step5: domain view 분리
