+++ src/main/kotlin/step3_4_5/controller/main.kt
@@ -1,7 +1,7 @@
-package step3_4_5
+package step3_4_5.controller
 
-import step3_4_5.racing.Car
-import step3_4_5.racing.CarRacing
+import step3_4_5.domain.racing.Car
+import step3_4_5.domain.racing.CarRacing
 import step3_4_5.view.InputView
 import step3_4_5.view.OutputView
 import step3_4_5.view.ResultView
